+++ src/utils/TableTransactions.js
@@ -1,0 +1,3 @@
+const tableHeaders = ['Date', 'Type', 'Category', 'Comment', 'Sum', 'Balance'];
+
+export default tableHeaders;
